requirements.txt
Pillow
windnd
